requirements.txt
streamlit
Pillow
pdf2image
python-dotenv
google-generativeai
